add static_cast<int32_t> for devID

+++ ais-bench_workload/tool/ais_bench/backend/base/include/Base/DeviceManager/DeviceManager.h
@@ -32,7 +32,7 @@ struct DeviceContext {
         IDLE = 0,  // idle status
         USING      // running status
     } devStatus = IDLE;
-    int32_t devId = DEFAULT_VALUE;
+    int32_t devId = static_cast<int32_t>(DEFAULT_VALUE);
 };
 
 class DeviceManager {
